Set the timestamp of files when they are created.
Keil has a feature to set the time of a file when the file is created/modified. (It may also set last access, etc, but I'm not sure and don't really care about accessed.)

For some reason, they built all of the plumbing to let the file time be set, but never actually implemented it; that's left as an exercise for the user.