Accommodate more .wav files, improvements to stats logging.
Mostly, this change is about making the .wav file playback work with more files. Some .wav files contain "chunk" types that weren't expected, with metadata that we don't care about. Those chunks can be arbitrary size, and can be between chunks that we *do* care about, so we can't simply read the first N bytes of the file. Now we read and validate the header, then read chunks until we find the "data" chunk with the audio bytes. Along the way we read the "fmt " chunk with audio properties.

Before this change, the wav files would play, but some of them had a "pop" at the start.

There is some new logging as well, that came out of debugging the the pops. The new logging should grow into a consistent set of log entries from which we can construct the play statistics for the dashboard (play/pause/stop/finish/seek, etc.)