Reconfigure the volume settings.
The volume previously was settable over a range of 10 or 11 values. The
loudest was well past distortion, and sometimes drew half an amp from the
battery (!), and could cause the codec to shut down due to overload (it
would reset at the end of the track). Several of the lower settings were
so faint that they could barely be heard; completely useless outside, and
of questionable utility indoors.

This change:
- Lowers the maximum volume to a value between the previous 8 and 9.
- Raises the minimum value to a value about the previous 4.
- Keeps the default (about) where it was (configurable through the
  control.def file.
- Spreads these over a range of '1' to '8', with '4' as the default.