Convert (ie, rename) main.c -> main.cpp.
This is the first baby step to converting the files to C++, at which point
we'll be able to use classes to provide namespacing variables.

There won't be much use for virtual functions, although audio.c can "play"
three different types of audio (.wav, .mp3, tones), and record another 1-1/2
(.wav, encrypted .wav), and is as likely a candidate as any. What I want
most is simply to gather data and its code together.

- Surround contents of tbook.h with
    # ifdef __cplusplus
    extern"C" {
    # endif
  . . .
    # ifdef __cplusplus
    }
    # endif